+++ tool/grind.dart
@@ -26,9 +26,6 @@ void _concatenateFiles(Directory src, File target) {
   String str = _traverse(src, '').map((s) => '  ${_toStr(s)}').join(',\n');
 
   target.writeAsStringSync("""
-// Copyright (c) 2014, Google Inc. Please see the AUTHORS file for details.
-// All rights reserved. Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
 const List<String> data = const [
 ${str}
 ];